requirements.txt
# Core
pandas
numpy
scikit-learn
matplotlib
seaborn

# Deep learning
tensorflow==2.17.0

# Evaluación y monitoreo
#evidently==0.4.26

# Utilidades
#jupyter
#ipykernel

#WHISPER OPEN AI
openai-whisper
torch
torchvision
torchaudio
numpy
tqdm
